refactor: refactored folder structure for helpers gf-382
