+++ apps/frontend/src/pages/access-management/libs/helpers/get-group-columns/get-group-columns.helper.tsx
@@ -1,8 +1,8 @@
 import { formatDate } from "~/libs/helpers/helpers.js";
 import { type TableColumn } from "~/libs/types/types.js";
 
-import { GroupMenu } from "../components/components.js";
-import { type GroupRow } from "../types/types.js";
+import { GroupMenu } from "../../components/components.js";
+import { type GroupRow } from "../../types/types.js";
 
 const getGroupColumns = (actions: {
 	onDelete: (groupId: number) => void;
