+++ apps/frontend/src/pages/access-management/libs/helpers/get-group-rows/get-group-rows.helper.tsx
@@ -1,6 +1,6 @@
 import { type GroupGetAllItemResponseDto } from "~/modules/groups/groups.js";
 
-import { type GroupRow } from "../types/types.js";
+import { type GroupRow } from "../../types/types.js";
 
 const getGroupRows = (groups: GroupGetAllItemResponseDto[]): GroupRow[] =>
 	groups.map((group) => ({
