+++ apps/frontend/src/pages/access-management/libs/helpers/helpers.ts
@@ -1,4 +1,4 @@
-export { getGroupColumns } from "./get-group-columns.helper.jsx";
-export { getGroupRows } from "./get-group-rows.helper.js";
-export { getUserColumns } from "./get-user-columns.helper.js";
-export { getUserRows } from "./get-user-rows.helper.js";
+export { getGroupColumns } from "./get-group-columns/get-group-columns.helper.jsx";
+export { getGroupRows } from "./get-group-rows/get-group-rows.helper.js";
+export { getUserColumns } from "./get-user-columns/get-user-columns.helper.js";
+export { getUserRows } from "./get-user-rows/get-user-rows.helper.js";
