+++ apps/frontend/src/pages/contributors/libs/helpers/helpers.ts
@@ -1,2 +1,2 @@
-export { getContributorColumns } from "./get-contributor-columns.helper.js";
-export { getContributorRows } from "./get-contributor-rows.helper.js";
+export { getContributorColumns } from "./get-contributor-columns/get-contributor-columns.helper.js";
+export { getContributorRows } from "./get-contributor-rows/get-contributor-rows.helper.js";
