+++ apps/frontend/src/pages/project-access-management/libs/helpers/filter-user-project-groups/filter-user-project-groups.helper.ts
