+++ apps/frontend/src/pages/project-access-management/libs/helpers/get-project-group-columns/get-project-group-columns.helper.tsx
@@ -1,8 +1,8 @@
 import { formatDate } from "~/libs/helpers/helpers.js";
 import { type TableColumn } from "~/libs/types/types.js";
 
-import { ProjectGroupMenu } from "../components/components.js";
-import { type ProjectGroupRow } from "../types/types.js";
+import { ProjectGroupMenu } from "../../components/components.js";
+import { type ProjectGroupRow } from "../../types/types.js";
 
 const getProjectGroupColumns = (actions: {
 	onDelete: (projectGroupId: number) => void;
