+++ apps/frontend/src/pages/project-access-management/libs/helpers/get-project-group-rows/get-project-group-rows.helper.tsx
@@ -1,6 +1,6 @@
 import { type ProjectGroupGetAllItemResponseDto } from "@git-fit/shared";
 
-import { type ProjectGroupRow } from "../types/types.js";
+import { type ProjectGroupRow } from "../../types/types.js";
 
 const getProjectGroupRows = (
 	projectGroups: ProjectGroupGetAllItemResponseDto[],
