+++ apps/frontend/src/pages/project-access-management/libs/helpers/get-user-rows/get-user-rows.helper.ts
@@ -1,6 +1,6 @@
 import { type UserGetAllItemResponseDto } from "~/modules/users/users.js";
 
-import { type UserRow } from "../types/types.js";
+import { type UserRow } from "../../types/types.js";
 
 const getUserRows = (users: UserGetAllItemResponseDto[]): UserRow[] =>
 	users.map((user) => ({
