+++ apps/frontend/src/pages/project-access-management/libs/helpers/helpers.ts
@@ -1,5 +1,5 @@
-export { filterUserProjectGroups } from "./filter-user-project-groups.helper.js";
-export { getProjectGroupColumns } from "./get-project-group-columns.helper.js";
-export { getProjectGroupRows } from "./get-project-group-rows.helper.js";
-export { getUserColumns } from "./get-user-columns.helper.js";
-export { getUserRows } from "./get-user-rows.helper.js";
+export { filterUserProjectGroups } from "./filter-user-project-groups/filter-user-project-groups.helper.js";
+export { getProjectGroupColumns } from "./get-project-group-columns/get-project-group-columns.helper.js";
+export { getProjectGroupRows } from "./get-project-group-rows/get-project-group-rows.helper.js";
+export { getUserColumns } from "./get-user-columns/get-user-columns.helper.js";
+export { getUserRows } from "./get-user-rows/get-user-rows.helper.js";
